requirements.txt
uvicorn
gunicorn
python-dotenv
redis
fastapi
qdrant_client
sqlacodegen
psycopg2-binary
sqlalchemy
twilio
llama-index
llama-index-vector-stores-qdrant
llama-index-agent-openai
llama-index-llms-anthropic
llama-index-llms-groq
python-multipart
aiosmtplib
livekit-api